Introduce `BaseProjectedMiniWorldRender` (which pulls out chunks of logic from `ProjectedMiniWorldRender`)
